+++ ignite/pkg/cosmosanalysis/app/testdata/modules/package_function/app.go
@@ -1,0 +1,23 @@
+package app
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	abci "github.com/tendermint/tendermint/abci/types"
+)
+
+var ModuleBasics = module.NewBasicManager(basicModules()...)
+
+type Foo struct{}
+
+func (Foo) Name() string {
+	return "foo"
+}
+
+func (Foo) BeginBlocker(sdk.Context, abci.RequestBeginBlock) abci.ResponseBeginBlock {
+	return abci.ResponseBeginBlock{}
+}
+
+func (Foo) EndBlocker(sdk.Context, abci.RequestEndBlock) abci.ResponseEndBlock {
+	return abci.ResponseEndBlock{}
+}