+++ ignite/pkg/cosmosanalysis/app/testdata/modules/registration_not_in_app_go/modules.go
@@ -1,0 +1,20 @@
+package app
+
+import (
+	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/bank"
+	"github.com/cosmos/cosmos-sdk/x/gov"
+	"github.com/cosmos/cosmos-sdk/x/staking"
+	foomodule "github.com/username/test/x/foo"
+)
+
+var ModuleBasics = module.NewBasicManager(
+	auth.AppModuleBasic{},
+	bank.AppModuleBasic{},
+	staking.AppModuleBasic{},
+	gov.NewAppModuleBasic([]govclient.ProposalHandler{
+		paramsclient.ProposalHandler,
+	}),
+	foomodule.AppModuleBasic{},
+)