+++ ignite/pkg/cosmosanalysis/module/testdata/planet/x/mars/keeper/grpc_query_my_query.go
@@ -1,0 +1,13 @@
+package keeper
+
+import (
+	"context"
+
+	"github.com/tendermint/planet/x/mars/types"
+)
+
+type Keeper struct{}
+
+func (k Keeper) MyQuery(goCtx context.Context, req *types.QueryMyQueryRequest) (*types.QueryMyQueryResponse, error) {
+	return nil, nil
+}