+++ ignite/templates/typed/singleton/stargate.go
@@ -98,16 +98,17 @@ const (
 
 func protoRPCModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := filepath.Join(opts.AppPath, "proto", opts.ModuleName, "query.proto")
+		path := filepath.Join(opts.AppPath, "proto", opts.AppName, opts.ModuleName, "query.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err
 		}
 
 		// Import the type
-		templateImport := `import "%s/%s.proto";
+		templateImport := `import "%s/%s/%s.proto";
 %s`
 		replacementImport := fmt.Sprintf(templateImport,
+			opts.AppName,
 			opts.ModuleName,
 			opts.TypeName.Snake,
 			typed.Placeholder,
@@ -171,17 +172,18 @@ func clientCliQueryModify(replacer placeholder.Replacer, opts *typed.Options) ge
 
 func genesisProtoModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := filepath.Join(opts.AppPath, "proto", opts.ModuleName, "genesis.proto")
+		path := filepath.Join(opts.AppPath, "proto", opts.AppName, opts.ModuleName, "genesis.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err
 		}
 
-		templateProtoImport := `import "%[2]v/%[3]v.proto";
+		templateProtoImport := `import "%[2]v/%[3]v/%[4]v.proto";
 %[1]v`
 		replacementProtoImport := fmt.Sprintf(
 			templateProtoImport,
 			typed.PlaceholderGenesisProtoImport,
+			opts.AppName,
 			opts.ModuleName,
 			opts.TypeName.Snake,
 		)
@@ -344,16 +346,17 @@ if found {
 
 func protoTxModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := filepath.Join(opts.AppPath, "proto", opts.ModuleName, "tx.proto")
+		path := filepath.Join(opts.AppPath, "proto", opts.AppName, opts.ModuleName, "tx.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err
 		}
 
 		// Import
-		templateImport := `import "%s/%s.proto";
+		templateImport := `import "%s/%s/%s.proto";
 %s`
 		replacementImport := fmt.Sprintf(templateImport,
+			opts.AppName,
 			opts.ModuleName,
 			opts.TypeName.Snake,
 			typed.PlaceholderProtoTxImport,
@@ -380,7 +383,7 @@ func protoTxModify(replacer placeholder.Replacer, opts *typed.Options) genny.Run
 		protoImports := opts.Fields.ProtoImports()
 		for _, f := range opts.Fields.Custom() {
 			protoImports = append(protoImports,
-				fmt.Sprintf("%[1]v/%[2]v.proto", opts.ModuleName, f),
+				fmt.Sprintf("%[1]v/%[2]v/%[3]v.proto", opts.AppName, opts.ModuleName, f),
 			)
 		}
 		for _, f := range protoImports {